Animation

+++ src/main/java/Animation/Animation.java
@@ -4,10 +4,10 @@ import javafx.scene.image.Image;
 
 public class Animation {
     private Image anim[][];
-    private Integer speed;
+    private Double speed;
     private Boolean loop;
 
-    public Animation(Image[][] animation, Integer speed, Boolean loop){
+    public Animation(Image[][] animation, Double speed, Boolean loop){
         this.anim = animation;
         this.speed = speed;
         this.loop = loop;
@@ -17,7 +17,7 @@ public class Animation {
         return anim;
     }
 
-    public Integer getSpeed(){
+    public Double getSpeed(){
         return speed;
     }
